close #5 close #2 more work for #7

+++ R/g_.R
@@ -25,7 +25,7 @@ NULL
 
 g_exp <- function(n, m) {
   neet::assert_neet(m, "numeric")
-  neet::assert_neet(n, "integer")
+  neet::assert_neet(n, "numeric")
 
   # Estimate parameters.
   lambda <- log(2) / m
@@ -39,7 +39,7 @@ g_exp <- function(n, m) {
 #' @export
 
 g_norm <- function(n, m, spread, spread_type = "iqr") {
-  neet::assert_neet(n, "integer")
+  neet::assert_neet(n, "numeric")
   neet::assert_neet(m, "numeric")
   neet::assert_neet(spread, "numeric")
   neet::assert_neet(spread_type, "character")
@@ -70,7 +70,7 @@ g_norm <- function(n, m, spread, spread_type = "iqr") {
 #' @export
 
 g_lnorm <- function(n, m, spread, spread_type = "iqr") {
-  neet::assert_neet(n, "integer")
+  neet::assert_neet(n, "numeric")
   neet::assert_neet(m, "numeric")
   neet::assert_neet(spread, "numeric")
   neet::assert_neet(spread_type, "character")
@@ -105,7 +105,7 @@ g_lnorm <- function(n, m, spread, spread_type = "iqr") {
 #' @export
 
 g_cauchy <- function(n, m, spread, spread_type = "iqr") {
-  neet::assert_neet(n, "integer")
+  neet::assert_neet(n, "numeric")
   neet::assert_neet(m, "numeric")
   neet::assert_neet(spread, "numeric")
   neet::assert_neet(spread_type, "character")
